Add accounts to state
